utils: rgn funcs to use uint32_t

+++ src/lib/utils/utils.cpp
@@ -1,12 +1,12 @@
 #include "utils.h"
 //#include "debug.h"
 
-unsigned long seed = 0;
+uint32_t seed = 0;
 
 // returns values between 0 and 0x7FFF
 // NB rngN depends on this output range, so if we change the
 // behaviour rngN will need updating
-long rng(void)
+uint32_t rng(void)
 {
     long m = 2147483648;
     long a = 214013;
@@ -15,7 +15,7 @@ long rng(void)
     return seed >> 16;
 }
 
-void rngSeed(long newSeed)
+void rngSeed(uint32_t newSeed)
 {
     seed = newSeed;
 }
@@ -23,29 +23,29 @@ void rngSeed(long newSeed)
 // returns 0<x<max where max <= 256
 // (actual upper limit is higher, but there is one and I haven't
 //  thought carefully about what it is)
-unsigned int rngN(unsigned int max)
+uint32_t rngN(uint32_t max)
 {
-    unsigned long x = rng();
-    unsigned int result = (x * max) / RNG_MAX;
+    uint32_t x = rng();
+    uint32_t result = (x * max) / RNG_MAX;
     return result;
 }
 
 // 0..255 returned
-long rng8Bit(void)
+uint32_t rng8Bit(void)
 {
     return rng() & 0b11111111;
 }
 
 // 0..31 returned
-long rng5Bit(void)
+uint32_t rng5Bit(void)
 {
     return rng() & 0b11111;
 }
 
 // 0..255 returned
-long rng0to2(void)
+uint32_t rng0to2(void)
 {
-    int randomNumber = rng() & 0b11;
+    uint32_t randomNumber = rng() & 0b11;
 
     while (randomNumber == 3)
     {
